Rename Throttle Condition related DTO names
